create a movie: main information ok
