+++ apps/festival/festival/src/app/dashboard/dashboard.component.html
@@ -5,7 +5,7 @@
         <mat-icon matListIcon svgIcon="dashboard"></mat-icon>
         <span matLine><b>Dashboard</b></span>
       </a>
-      <a mat-list-item routerLink="title" routerLinkActive="active">
+      <a mat-list-item routerLink="title" routerLinkActive="active" test-id="title">
         <mat-icon matListIcon svgIcon="video_library"></mat-icon>
         <span matLine><b>My Titles</b></span>
       </a>
