+++ libs/testing/cypress/browser/src/lib/support.ts
@@ -26,7 +26,7 @@ export function get(selector: string) {
   return cy.get(`[test-id="${selector}"]`);
 }
 
-function getAllStartingWith(selector: string) {
+export function getAllStartingWith(selector: string) {
   return cy.get(`[test-id^="${selector}"]`);
 }
