+++ libs/ui/src/lib/dashboard/components/empty-movie/empty-movie.component.html
@@ -1,8 +1,8 @@
-<dashboard-no-title></dashboard-no-title>
+<dashboard-no-title test-id="no-title"></dashboard-no-title>
 <article fxLayout="row" fxLayout.lt-md="column" fxLayoutAlign="center" fxLayoutGap="24px">
   <div class="surface" fxLayoutalign="center center">
     <img asset="add_title.svg" alt="Add title">
-    <a routerLink="lobby" mat-button color="primary">Add one Title</a>
+    <a routerLink="lobby" mat-button color="primary" test-id="add-title">Add one Title</a>
   </div>
   <div class="surface">
     <img asset="add_files.svg" alt="Upload files">
